requirements.txt
aiogram==3.13.1
openai>=1.40.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
